Move createIndexError tests to unit/
